* subversion/libsvn_client/shelf.c (shelf_name_from_filename): Use consistent data types.
git-svn-id: https://svn.apache.org/repos/asf/subversion/trunk@1852384 13f79535-47bb-0310-9956-ffa450edef68

+++ subversion/libsvn_client/shelf.c
@@ -100,7 +100,7 @@ shelf_name_from_filename(char **name,
 {
   size_t len = strlen(filename);
   static const char suffix[] = ".current";
-  int suffix_len = sizeof(suffix) - 1;
+  size_t suffix_len = sizeof(suffix) - 1;
 
   if (len > suffix_len && strcmp(filename + len - suffix_len, suffix) == 0)
     {
